Moved the DPKGPackageFactory into the dpkg package
