+++ services/package-manager/src/main/java/org/openthinclient/util/dpkg/DPKGPackageFactory.java
@@ -18,7 +18,7 @@
  * this program; if not, write to the Free Software Foundation, Inc., 59 Temple
  * Place - Suite 330, Boston, MA 02111-1307, USA.
  *******************************************************************************/
-package org.openthinclient.pkgmgr;
+package org.openthinclient.util.dpkg;
 
 import java.io.BufferedReader;
 import java.io.File;
